reorganize eks test file structure
