+++ test/vllm/eks/test_artifacts/eks-cluster.yaml
