+++ test/vllm/eks/test_artifacts/fsx-lustre-pvc.yaml
