+++ test/vllm/eks/test_artifacts/fsx-storage-class.yaml
