+++ test/vllm/eks/test_artifacts/large-model-nodegroup.yaml
