+++ test/vllm/eks/test_artifacts/vllm-deepseek-32b-lws-ingress.yaml
